requirements.txt
paho-mqtt
git+https://github.com/uintptr/jsonconfig.git
